requirements.txt
PyMuPDF>=1.24
numpy>=1.26
opencv-python-headless>=4.10
shapely>=2.0
pydantic>=2.7
Pillow>=10.0
pytest>=8.0
pytesseract>=0.3

